+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 - Send microsecond precision timestamps. ([#3613](https://github.com/getsentry/relay/pull/3613))
 - Extract additional user fields for spans. ([#3599](https://github.com/getsentry/relay/pull/3599))
 - Disable `db.redis` span metrics extraction. ([#3600](https://github.com/getsentry/relay/pull/3600))
+- Extract status for spans. ([#3606](https://github.com/getsentry/relay/pull/3606))
 
 ## 24.4.2